requirements.txt
allure_python_commons==2.8.18
pytest==6.1.1
requests==2.24.0
xlrd==1.2.0
PyYAML==5.3.1
python-dateutil==2.8.1
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0
